be: drop the jobs grouping, generate-types/process-set are top-level commands
Everything under be/cmd/ is already a command by virtue of living there -
a shared "jobs" parent command added an extra layer of indirection
(be jobs generate-types / be jobs process-set) for no real benefit.
generate-types and process-set are each their own top-level cobra command
now (be generate-types, be process-set), split into their own files
(generate_types.go, process_set.go) named for what they represent rather
than a general jobs.go someone would need to open to find out.
